requirements.txt
ccxt==2.2.79
config==0.5.1
Flask==2.2.2
pandas==1.5.2
python-dotenv==1.0.0
python_telegram_bot==13.8
requests==2.28.1
telegram==0.0.1
websocket_client==1.5.1
